First Semester/Data Intensive/Exercise_3/base-docker-app/obj_detection_container-vanilla/requirements.txt
Flask
Pillow
tensorflow
numpy
tensorflow-hub